split addr types into modules
