+++ kernel/src/mm/addr.rs
@@ -1,0 +1,9 @@
+mod kernel_virt_addr;
+mod phys_addr;
+mod user_ptr;
+mod virt_addr;
+
+pub use kernel_virt_addr::*;
+pub use phys_addr::*;
+pub use user_ptr::*;
+pub use virt_addr::*;
